Crate ipld_blockstore

Modules

resolve

Structs

BSStats

BufferedBlockStore
Wrapper around BlockStore to limit and have control over when values are written. This type is not threadsafe and can only be used in synchronous contexts.

TrackingBlockStore
Wrapper around BlockStore to tracking reads and writes for verification. This struct should only be used for testing.

Traits

BlockStore
Wrapper for database to handle inserting and retrieving ipld data with Cids