Crate ipld_blockstore

Modules

resolve
This module is used for resolving Cids and Ipld recursively. This is generally only needed for testing because links should generally not be collapsed to generate a singular data structure, or this would lead to ambiguity of the data.

Structs

BufferedBlockStore
Wrapper around BlockStore to limit and have control over when values are written. This type is not threadsafe and can only be used in synchronous contexts.

Traits

BlockStore
Wrapper for database to handle inserting and retrieving ipld data with Cids