Trait ipld_blockstore::BlockStore

pub trait BlockStore: Store {
    fn get_bytes(&self, cid: &Cid) -> Result<Option<Vec<u8>>, Box<dyn StdError>> { ... }

    fn get<T>(&self, cid: &Cid) -> Result<Option<T>, Box<dyn StdError>>
    where
        T: DeserializeOwned,
    { ... }

    fn put<S>(&self, obj: &S, code: Code) -> Result<Cid, Box<dyn StdError>>
    where
        S: Serialize,
    { ... }

    fn put_raw(
        &self, 
        bytes: Vec<u8>, 
        code: Code
    ) -> Result<Cid, Box<dyn StdError>> { ... }

    fn bulk_put<'a, S, V>(
        &self, 
        values: V, 
        code: Code
    ) -> Result<Vec<Cid>, Box<dyn StdError>>
    where
        S: Serialize + 'a,
        V: IntoIterator<Item = &'a S>,
    { ... }
}

Wrapper for database to handle inserting and retrieving ipld data with Cids

Provided methods

fn get_bytes(&self, cid: &Cid) -> Result<Option<Vec<u8>>, Box<dyn StdError>>

Get bytes from block store by Cid.

fn get<T>(&self, cid: &Cid) -> Result<Option<T>, Box<dyn StdError>> where
T: DeserializeOwned,

Get typed object from block store by Cid.

fn put<S>(&self, obj: &S, code: Code) -> Result<Cid, Box<dyn StdError>> where
S: Serialize,

Put an object in the block store and return the Cid identifier.

fn put_raw(&self, bytes: Vec<u8>, code: Code) -> Result<Cid, Box<dyn StdError>>

Put raw bytes in the block store and return the Cid identifier.

fn bulk_put<'a, S, V>(
&self,
values: V,
code: Code
) -> Result<Vec<Cid>, Box<dyn StdError>> where
S: Serialize + 'a,
V: IntoIterator<Item = &'a S>,

Batch put cbor objects into blockstore and returns vector of Cids

Loading content...