Trait ipld_blockstore::BlockStore

pub trait BlockStore: Store {
    fn get_bytes(&self, cid: &Cid) -> Result<Option<Vec<u8>>, Box<dyn StdError>> { ... }

    fn get<T>(&self, cid: &Cid) -> Result<Option<T>, Box<dyn StdError>>
    where
        T: DeserializeOwned,
    { ... }

    fn put<S, T>(&self, obj: &S, hash: T) -> Result<Cid, Box<dyn StdError>>
    where
        S: Serialize,
        T: MultihashDigest,
    { ... }
}

Wrapper for database to handle inserting and retrieving ipld data with Cids

Provided methods

fn get_bytes(&self, cid: &Cid) -> Result<Option<Vec<u8>>, Box<dyn StdError>>

Get bytes from block store by Cid

fn get<T>(&self, cid: &Cid) -> Result<Option<T>, Box<dyn StdError>> where
T: DeserializeOwned,

Get typed object from block store by Cid

fn put<S, T>(&self, obj: &S, hash: T) -> Result<Cid, Box<dyn StdError>> where
S: Serialize,
T: MultihashDigest,

Put an object in the block store and return the Cid identifier

Loading content...